+++ sessionize-base/src/main/kotlin/info/metadude/kotlin/library/sessionize/adapters/ZonedDateTimeAdapter.kt
@@ -3,28 +3,28 @@ package info.metadude.kotlin.library.sessionize.adapters
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.ToJson
 import org.threeten.bp.DateTimeException
-import org.threeten.bp.LocalDateTime
+import org.threeten.bp.ZonedDateTime
 import org.threeten.bp.format.DateTimeFormatter
 import org.threeten.bp.format.DateTimeParseException
 
-class LocalDateTimeAdapter {
+class ZonedDateTimeAdapter {
 
-    private val dateTimeFormatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME!!
+    private val dateTimeFormatter = DateTimeFormatter.ISO_OFFSET_DATE_TIME!!
 
     @FromJson
     fun fromJson(text: String?) = text?.let {
         try {
-            LocalDateTime.parse(text, dateTimeFormatter)
-        } catch (e: DateTimeParseException) {
+            ZonedDateTime.parse(text, dateTimeFormatter)
+        } catch (_: DateTimeParseException) {
             null
         }
     }
 
     @ToJson
-    fun toJson(dateTime: LocalDateTime?) = dateTime?.let {
+    fun toJson(dateTime: ZonedDateTime?) = dateTime?.let {
         try {
             it.format(dateTimeFormatter)
-        } catch (e: DateTimeException) {
+        } catch (_: DateTimeException) {
             null
         }
     }