+++ sessionize-base/src/main/kotlin/info/metadude/kotlin/library/sessionize/gridtable/models/ConferenceDay.kt
@@ -1,12 +1,12 @@
 package info.metadude.kotlin.library.sessionize.gridtable.models
 
 import com.squareup.moshi.JsonClass
-import org.threeten.bp.LocalDateTime
+import org.threeten.bp.ZonedDateTime
 
 @JsonClass(generateAdapter = true)
 data class ConferenceDay(
 
-        val date: LocalDateTime = LocalDateTime.MIN,
+        val date: ZonedDateTime? = null,
         val rooms: List<Room> = emptyList(),
         val timeSlots: List<TimeSlot> = emptyList()